2_疲劳度检测/code/Blink_detection/requirements.txt
opencv-python~=4.5.5.64
numpy~=1.21.6
tensorflow-gpu~=1.13.1
keras~=2.1.5
matplotlib~=3.5.2
pillow~=9.1.0